// src/modules/kradiator/typescript/types/kradSiteCsvRadiator.type.ts
export type KradSiteCsvRadiator = {
    name: string	
    description: string
    columns:string	
    interaxle: string	
    sections: string

    price: string	

    height: string	
    depth: string	
    length: string	

    weight: string	
    volume: string	
    dt70: string	
    dt60: string
    
    wall_consoles_name: string	
    wall_consoles_image: string	
}